requirements.txt
aiohttp>=3.8.0
aiogram>=3.0.0
apscheduler>=3.8.1
beautifulsoup4>=4.9.3
python-dotenv==1.0.0
validators>=0.18.2
numpy==1.26.3
matplotlib>=3.4.3
